+++ test/neighbor_test.rb
@@ -280,6 +280,24 @@ class NeighborTest < Minitest::Test
     assert_elements_in_delta [6, 4], result.map(&:neighbor_distance)
   end
 
+  def test_hamming
+    Item.create!(id: 1, binary_embedding: "000")
+    Item.create!(id: 2, binary_embedding: "101")
+    Item.create!(id: 3, binary_embedding: "111")
+    result = Item.nearest_neighbors(:binary_embedding, "101", distance: "hamming").first(5)
+    assert_equal [2, 3, 1], result.map(&:id)
+    assert_elements_in_delta [0, 1, 2], result.map(&:neighbor_distance)
+  end
+
+  def test_jaccard
+    Item.create!(id: 1, binary_embedding: "000")
+    Item.create!(id: 2, binary_embedding: "101")
+    Item.create!(id: 3, binary_embedding: "111")
+    result = Item.nearest_neighbors(:binary_embedding, "100", distance: "jaccard").first(5)
+    assert_equal [2, 3, 1], result.map(&:id)
+    assert_elements_in_delta [0.5, 2/3.0, 1], result.map(&:neighbor_distance)
+  end
+
   def create_items(cls, attribute = :embedding)
     vectors = [
       [1, 1, 1],