+++ test/support/cube.rb
@@ -5,11 +5,12 @@ ActiveRecord::Schema.define do
     t.cube :embedding
     t.cube :neighbor_vector
     t.cube :factors
+    t.bit :binary_embedding, limit: 3
   end
 end
 
 class Item < ActiveRecord::Base
-  has_neighbors :embedding
+  has_neighbors :embedding, :binary_embedding
   has_neighbors
 end