+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.3.3 (unreleased)
 
 - Added support for `halfvec` type
-- Added support for `taxicab` distance with `vector` extension
+- Added support for `taxicab`, `hamming`, and `jaccard` distances with `vector` extension
 - Fixed issue with `nearest_neighbors` scope overriding `select` values
 
 ## 0.3.2 (2023-12-12)